+++ types/tab.d.ts
@@ -1,3 +1,0 @@
-export function setupTabs(): void;
-export function prepareTabFonts(): void;
-export function getBlitzkriegTabIndex(): sc.OPTION_CATEGORY;